+++ src/RecoveryServices/RecoveryServices.SiteRecovery.Autorest/custom/Cmdlets/Policy/Remove-AzRecoveryServicesReplicationPolicy.ps1
@@ -1,6 +1,6 @@
 function Remove-AzRecoveryServicesReplicationPolicy
 {
-	[OutputType([Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Models.Api20230201.IJob])]
+	[OutputType([System.Boolean])]
     [CmdletBinding(PositionalBinding=$false, SupportsShouldProcess)]
     [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Description('Removes a given replication policy in a given recovery services vault')]
 
@@ -11,14 +11,17 @@ function Remove-AzRecoveryServicesReplicationPolicy
         ${SubscriptionId},
 
         [Parameter(Mandatory, HelpMessage='The name of the resource group where the recovery services vault is present.')]
+        [ValidateNotNullOrEmpty()]
         [System.String]
         ${ResourceGroupName},
 
         [Parameter(Mandatory, HelpMessage='The name of the recovery services vault.')]
+        [ValidateNotNullOrEmpty()]
         [System.String]
         ${ResourceName},
 
         [Parameter(Mandatory, HelpMessage='Specific replication policy object.')]
+        [ValidateNotNull()]
         [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Models.Api20230201.IPolicy]
         ${Policy},