+++ src/RecoveryServices/RecoveryServices.SiteRecovery.Autorest/custom/Cmdlets/ProtectionContainer/Get-AzRecoveryServicesReplicationProtectionContainer.ps1
@@ -1,0 +1,127 @@
+
+# ----------------------------------------------------------------------------------
+#
+# Copyright Microsoft Corporation
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ----------------------------------------------------------------------------------
+
+<#
+.Synopsis
+Gets the details of a protection container.
+.Description
+Gets the details of a protection container.
+
+.Outputs
+Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Models.Api20230201.IProtectionContainer
+.Link
+https://docs.microsoft.com/powershell/module/az.recoveryservices/get-azrecoveryservicesreplicationprotectioncontainer
+#>
+function Get-AzRecoveryServicesReplicationProtectionContainer {
+    [OutputType([Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Models.Api20230201.IProtectionContainer])]
+    [CmdletBinding(PositionalBinding=$false)]
+    param(
+        [Parameter(ParameterSetName='List',Mandatory)]
+        [Parameter(ParameterSetName='Get',Mandatory)]
+        [ValidateNotNull()]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Path')]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Models.Api20230201.IFabric]
+        # Fabric Object.
+        ${Fabric},
+
+        [Parameter(Mandatory)]
+        [ValidateNotNullOrEmpty()]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Path')]
+        [System.String]
+        # The name of the resource group where the recovery services vault is present.
+        ${ResourceGroupName},
+
+        [Parameter(Mandatory)]
+        [ValidateNotNullOrEmpty()]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Path')]
+        [System.String]
+        # The name of the recovery services vault.
+        ${ResourceName},
+
+        [Parameter(ParameterSetName='Get', Mandatory)]
+        [ValidateNotNullOrEmpty()]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Path')]
+        [System.String]
+        # Protection container name.
+        ${ProtectionContainerName},
+
+        [Parameter()]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Path')]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Runtime.DefaultInfo(Script='(Get-AzContext).Subscription.Id')]
+        [System.String[]]
+        # The subscription Id.
+        ${SubscriptionId},
+
+        [Parameter()]
+        [Alias('AzureRMContext', 'AzureCredential')]
+        [ValidateNotNull()]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Azure')]
+        [System.Management.Automation.PSObject]
+        # The credentials, account, tenant, and subscription used for communication with Azure.
+        ${DefaultProfile},
+
+        [Parameter(DontShow)]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Runtime')]
+        [System.Management.Automation.SwitchParameter]
+        # Wait for .NET debugger to attach
+        ${Break},
+
+        [Parameter(DontShow)]
+        [ValidateNotNull()]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Runtime')]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Runtime.SendAsyncStep[]]
+        # SendAsync Pipeline Steps to be appended to the front of the pipeline
+        ${HttpPipelineAppend},
+
+        [Parameter(DontShow)]
+        [ValidateNotNull()]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Runtime')]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Runtime.SendAsyncStep[]]
+        # SendAsync Pipeline Steps to be prepended to the front of the pipeline
+        ${HttpPipelinePrepend},
+
+        [Parameter(DontShow)]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Runtime')]
+        [System.Uri]
+        # The URI for the proxy server to use
+        ${Proxy},
+
+        [Parameter(DontShow)]
+        [ValidateNotNull()]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Runtime')]
+        [System.Management.Automation.PSCredential]
+        # Credentials for a proxy server to use for the remote call
+        ${ProxyCredential},
+
+        [Parameter(DontShow)]
+        [Microsoft.Azure.PowerShell.Cmdlets.RecoveryServices.Category('Runtime')]
+        [System.Management.Automation.SwitchParameter]
+        # Use the default credentials for the proxy
+        ${ProxyUseDefaultCredentials}
+    )
+
+    process {
+        try {
+            $fabricName = $Fabric.Name
+        
+            $null = $PSBoundParameters.Remove("Fabric")
+            $null = $PSBoundParameters.Add("FabricName", $fabricName)
+
+            return Az.RecoveryServices.internal\Get-AzRecoveryServicesReplicationProtectionContainer @PSBoundParameters
+        } catch {
+            throw
+        }
+    }
+}