+++ src/RecoveryServices/RecoveryServices.SiteRecovery.Autorest/test/A2A/Get-AzRecoveryServicesReplicationFabric.Tests.ps1
@@ -1,0 +1,27 @@
+if(($null -eq $TestName) -or ($TestName -contains 'Get-AzRecoveryServicesReplicationFabric'))
+{
+  $loadEnvPath = Join-Path $PSScriptRoot 'loadEnv.ps1'
+  if (-Not (Test-Path -Path $loadEnvPath)) {
+      $loadEnvPath = Join-Path $PSScriptRoot '..\loadEnv.ps1'
+  }
+  . ($loadEnvPath)
+  $TestRecordingFile = Join-Path $PSScriptRoot 'Get-AzRecoveryServicesReplicationFabric.Recording.json'
+  $currentPath = $PSScriptRoot
+  while(-not $mockingPath) {
+      $mockingPath = Get-ChildItem -Path $currentPath -Recurse -Include 'HttpPipelineMocking.ps1' -File
+      $currentPath = Split-Path -Path $currentPath -Parent
+  }
+  . ($mockingPath | Select-Object -First 1).FullName
+}
+
+Describe 'Get-AzRecoveryServicesReplicationFabric' {
+    It 'List' {
+        $output = Get-AzRecoveryServicesReplicationFabric -ResourceName $env.a2aVaultName -ResourceGroupName $env.a2aResourceGroupName -SubscriptionId $env.a2aSubscriptionId
+        $output.Count | Should Not BeNullOrEmpty
+    }
+
+    It 'Get' {
+        $output = Get-AzRecoveryServicesReplicationFabric -ResourceName $env.a2aVaultName -ResourceGroupName $env.a2aResourceGroupName -SubscriptionId $env.a2aSubscriptionId -FabricName $env.a2aFabricName
+        $output.Count | Should Not BeNullOrEmpty
+    }
+}